http2: fix Transport cancelation problems
The Transport code wasn't always sending a RST_STREAM to the server
in a few cases, and was also forgetting to flush its buffer.

I believe each part of this change was part of making golang/go#13556
not flaky. No additional tests at the moment;
TestTransportAndServerSharedBodyRace_h2 is pretty comprehensive as-is,
at least for an integration test. I should probably add some
finer-grained tests in the future.

Updates golang/go#13556

Change-Id: Ie824202b20ba6ee85b20edeeee86ec977e6daeb2
Reviewed-on: https://go-review.googlesource.com/18288
Reviewed-by: Andrew Gerrand <adg@golang.org>